Working on sliding window analysis and storage in JSON file

+++ src/Database/Entry.hpp
@@ -40,7 +40,7 @@ public:
     explicit Entry(string path);
 
     string path;
-    vector<float> markers;
+    vector<int> markers;
     vector<float> centroid;
     vector<float> flatness;
 };
